server/user_authorized_keys: Do not hardcode homedir
Previously, the ssh settings directory was hardcoded as
`/home/{user}/.ssh`, but this is not always in correct (e.g. the root
user typically has `/root`, and administrators can configure any
directory as a user's home directory).

This commit replaces the hardcoded default with a lookup using the Home
fact, which does an actual lookup (via the ~user shell feature, which
does getpwnam() under the hood).

This commit also removes the trailing / from the directory, as it is
added wherever needed anyway.