Cleanup and fix typing issues with latest mypy
Mainly facts - the changes to the `FactBase` types are breaking but not
any functionality. For typing fact classes to pass they must implement
the `command` & `requires_command` attributes as class methods, not
plain strings.

This change hugely simplifies the typing logic here (method or string?)
which the latest (1.11) mypy raised issues for. The underlying logic
remains unchanged and string attributes will continue to work for the
foreseeable future.